requirements.txt
pandas
requests
tqdm
binance-connector